Change the way assignees are specified.
Now you prefix a line with "username:" instead of USERNAME.  You also define
the list of users in the 'users' file instead of having a hardcoded list in
sched.py.  And each user can have aliases: as in /etc/hosts, the first word
is the 'primary' name for that user, the the others are aliases.  The
primary is the name schedulator actually prints out.