Support #date lines.
This lets you set the startdate for a particular person or for everyone.

In theory, inserting a line like this in the middle of a file would let you
fastforward the date for a person at that point, but that doesn't actually
work because we handle #date during parsing, even though the dates
themselves are calculated *after* that.