web: colour odd/even rows differently.
Actually, we do them in groups of two, using a trick I learned from
StarQuery.  This results in fewer horizontal lines but still makes it easy
to follow the grid with your eyes (since there's an edge along *either* the
top or the bottom of each task).

This looks slightly odd when a particular task occupies more than a single
line, however.