On the [All] tab, show *all* the users assigned to each task.
Some tasks have subtasks that are assigned to other people, and for some,
the parent task is assigned to nobody and only the subtasks have assignees.

Before, the parent task's assignee is the only one that was shown.  Now we
show the parent task's owner *and* the owners of any subtasks, which is more
accurate.