Import our tasks as multipart/digest instead of mbox.
A maildir full of mboxes sounded clever, but it actually wasn't that useful.
mutt can handle multipart/digest, which means you can open a single message
that contains a bunch of other messages - exactly what we want in this case.

Let's still (incorrectly, but oh well) start the message with a "From " line
so it can *also* be opened as a one-message mbox.  This doesn't seem to
disturb mutt and it means that our 'm ####' command still works.