Add a 'bog view' command to replace the silly 'm' script.
It views the given bog using mutt.  If your cwd is a subfolder of your bog
dir (ie. a maildir), it opens that folder in mutt.  If your cwd is anywhere
else, it opens the folder list in mutt.

And if you just use 'bog ####' (ie. just a bog number), or simply 'bog' by
itself, it defaults to 'view' for convenience.  I suspect that'll be 99% of
the use cases.  And if not, well, we can always change it :)