instaweb: concatenate tasks from a Maildir-style setup.
The current implementation of this is extremely evil, but does allow some
basic experimentation.

We also look for a main.sched in the root of the BOG_DIR, which works like
the old standalone test.sched schedule used to work.  That files comes
first, followed by the milestone folders in (currently) alphabetical order.

The example/ folder is an example of this setup.  'bog init' also now
creates an example empty Maildir to play with.