Replace table-based schedulator layout with ul/li/div-based layout.
Personally, I prefer the concept of table layouts for a lot of things.  But
nesting table rows into a hierarchy is just totally insane.  So let's use
nested lists instead.

This simplifies the css and the javascript, plus jquery's animations no
longer look jittery.

Bonus points: we now colour-code each Person in the schedulator display so
it's easy to see a particular person's task in the combined list.