cmd/web: don't cause conflicts when saving multiple times in the same form.
The early patch to make committing more careful by remembering the parent
commitid didn't *update* the parent commitid when the editor page saved
multiple times without being reloaded... because I forgot that would ever
happen.

The server now sends back the commitid and the client saves it for the next
commit.

Also, remove the old behaviour of saving every five seconds.  It was maybe
helpful back when we were using mobwrite, but not so important anymore.